+++ CHANGELOG.rst
@@ -71,6 +71,8 @@ Misc new features:
     DeprecationWarning: getargs: The 'u' format is deprecated. Use 'U' instead.
 - Added ``timeout`` and ``hard=True/False`` flags to `pysys.process.Process.stop`. Also added logic on Linux which will 
   automatically attempt a SIGKILL if the SIGTERM times out (though will still raise an exception in this case). 
+- Added ``closeStdinAfterWrite`` parameter to `pysys.process.Process.write` which can be used for child processes that 
+  wait for End Of File before completing. 
 
 Fixes: